Replace nonfree Aller Bold font with Fira Bold
It seems to me that the font used so far for HUD and in-game texts
was Aller Bold, which is proprietary: https://www.fontsquirrel.com/fonts/aller

I've replaced it by Fira Sans, which looks quite similar and is
distributed under the terms of the SIL Open Font License 1.0.

Part of #11.

Also fixes issues with the old BitMapFont being corrupted by recent
Godot - it's safer to just keep the TTF and use DynamicFont.

Note that I'm removing the reference to the ArmWrestler font from
the README. I don't see where it would be in use currently.